Toaster Added
Toaster added, that it Toaster.jsx and ToasterContext.jsx. These are in own files and App.jsx is now much cleaner as well. Will now do styling of the toasts for success of submit button. This will be done by also passing a type of the toast (and text as well of course).